Distributed system working, need to check final avg

+++ custom_msg/flatbuff/custom_ser.h
@@ -15,7 +15,7 @@ public:
    * @param value
    * @param buffer to reference during sending
    */
-  static void ser_msg(std::string msg, uint8_t id, int32_t value,
+  static void ser_msg(std::string msg, uint8_t id, float value,
                       flatbuffers::FlatBufferBuilder *builder, uint8_t *buf,
                       uint32_t *size) {
 
@@ -47,7 +47,7 @@ public:
    * @param value
    */
   static bool deser_msg(const uint8_t *buf, std::string &msg, uint8_t &id,
-                        int32_t &value) {
+                        float &value) {
 
     // Extract the message size from the buffer
     uint32_t size;
